Update page indicator with cuurent image cell to be same index
