+++ Nexus-Store/Scenes/ProductDetails/CustomNibCell/ProductImageCell/ProductImageCell.swift
@@ -17,7 +17,6 @@ class ProductImageCell: UICollectionViewCell {
     
     @IBOutlet weak var productImage: UIImageView!
     
-    @IBOutlet weak var EntireSelectedImage: UIPageControl!
     
     override func awakeFromNib() {
         super.awakeFromNib()
@@ -26,24 +25,6 @@ class ProductImageCell: UICollectionViewCell {
     }
 
 }
-
-////MARK: - Edit UI UIImageView
-//extension UIImageView {
-//
-//    func gradientImagetoBottom(){
-//        let gradient1 = CAGradientLayer()
-//        gradient1.startPoint = CGPoint(x: 0, y: 0)
-//        gradient1.endPoint = CGPoint(x: 0, y: 1)
-//        let whiteColor = UIColor.red
-//        gradient1.colors = [whiteColor.withAlphaComponent(1.0).cgColor, whiteColor.withAlphaComponent(1.0).cgColor, whiteColor.withAlphaComponent(0).cgColor]
-//        gradient1.locations = [NSNumber(value: 0), NSNumber(value: 0.5), NSNumber(value: 1)]
-//        gradient1.frame = self.frame
-//        self.layer.mask = gradient1
-//
-//    }
-//
-//}
-
 
 
 //MARK: - Edit UI UIImageView
